Wireframe of the first acceptance test.
*Given* that the Player has been started
*And* that the Casino has been started,
*And* that the Casino has a single Dealer named "Joe"
*When* the Player enters a Casino,
*Then* the Player should see a list of Dealers containing a Dealer named "Joe".

So far, all that really exists is a PlayerDriver that will drive the Player UI
using WindowLicker, and a CasinoManager that will connect to the Casino
management interface to create a Dealer and wait for the player to connect.

This seems to imply that the Casino server needs a management interface that can
add new Dealers and list the Dealers that exist.