Added windowlicker jars and a script to install them into your local maven repo.
The script can actually be used for any prebuilt jars we need to deal with. Just
put them in the prebuilt directory in a subdirectory named for the desired
groupId. The artifactId will be derived from the filename; "-DEV" and ".jar" will be
stripped from the basename of the file. Version will be set to "0.0.0-PREBUILT",
so if in the future an official Maven artifact becomes available, the maven
dependency search algorithm should work correctly. I've not confirmed that this
will work with maven-release-plugin, but I think it should.