We're now starting to play with connecting to the server.
In order to get this first acceptance test up and running, we need to control
the server through the management interface. We now attempt to do that, but we
fail, since the management interface isn't implemented yet.

I'm starting to feel that there's at least one new module here: the management
client, and possibly a second one with entities such as Dealers, Players, etc.
that would be shared between the Player client, the CasinoManager client and the
Casino server. I don't know if the management client needs a UI yet, but we
definitely are using it in the acceptance tests.