We can now "hire" a dealer through the manager interface.
Not really. It's actually just stubbed out enough for us to make forward
progress on the acceptance test for iteration 1. But we've got some more unit
tests for Dealer covering generating a text representation and comparing
Dealers, and a web.xml that will actually get the server started up in Jersey.